+++ packages/site/src/componentList.ts
@@ -54,13 +54,6 @@ export const componentList = [
     additionalMatches: ["Search", "Typeahead", "Suggest"],
   },
   {
-    title: "Autocomplete (v2)",
-    to: "/components/Autocomplete (v2)",
-    imageURL: "/Autocomplete.png",
-    sections: ["Forms & Inputs"],
-    additionalMatches: ["Search", "Typeahead", "Suggest", "Rebuilt"],
-  },
-  {
     title: "AutoLink",
     to: "/components/AutoLink",
     imageURL: "/AutoLink.png",